Always build Wasp project before deployment
